+++ TombEngine/Game/camera.h
@@ -96,7 +96,7 @@ void LookAt(CAMERA_INFO* cam, short roll);
 void AlterFOV(short value, bool store = true);
 short GetCurrentFOV();
 void AlterRoll(short value);
-float GetCurrentRoll();
+short GetCurrentRoll();
 void InitializeCamera();
 void MoveCamera(GameVector* ideal, int speed, bool force = false);
 void ChaseCamera(ItemInfo* item);